Reduced CFinalCommitment::quorumIndex from 32 to 16 bits

+++ src/llmq/commitment.h
@@ -28,7 +28,7 @@ public:
     uint16_t nVersion{CURRENT_VERSION};
     Consensus::LLMQType llmqType{Consensus::LLMQType::LLMQ_NONE};
     uint256 quorumHash;
-    int quorumIndex;
+    int16_t quorumIndex;
     std::vector<bool> signers;
     std::vector<bool> validMembers;
 
@@ -64,7 +64,7 @@ public:
                 obj.quorumHash
                 );
 
-        int quorumIndex = 0;
+        int16_t quorumIndex = 0;
         SER_WRITE(obj, quorumIndex = obj.quorumIndex);
         if (obj.nVersion == CFinalCommitment::INDEXED_QUORUM_VERSION) {
             READWRITE(quorumIndex);
